backend/requirements.txt
# Django Framework
Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
django-environ==0.11.2

# Datenbank Treiber
psycopg2-binary==2.9.10

# WSGI für Production


# Utilities
PyJWT==2.9.0
itsdangerous==2.2.0
sqlparse==0.5.3
asgiref==3.8.1
packaging==25.0
tzdata==2025.2
python-dotenv
Pillow
django-import-export
xhtml2pdf
dj-database-url

